Update register page to use leagues instead of registration leagues
Instead of signing up just for a night, you are actually signing up for
a specific level as well. Also, the system will enforce a cap on
registrations for each league so that a team cannot sign up for a
given league if it is already full. The cap is met simply by
registration at this point; payment is not required.

We should add something in the fine print saying that your payment
ensures your spot, so we can remove or change a team's registration if
necessary to open up a spot in a previously full division.